Add Spring Boot

See merge request ddrrdd/java!18

+++ trello-clone/src/main/resources/db/migration/V16__add_checklists_table.sql
@@ -3,7 +3,7 @@ create table checklists
     id       uuid primary key,
     name     varchar(60),
 
-    card_id  uuid     not null references cards (id),
-    position smallint not null,
+    card_id  uuid not null references cards (id),
+    position int4 not null,
     unique (card_id, position)
 )
